QEMU: various fixes regarding Thumb disassembly
Quirk: if CPU is in ARM mode, the disassembler looks at the Thumb bit of the requested address.
If CPU is in Thumb mode, the disassembler ignores the Thumb bit from the address, and always disassembles as Thumb.
This inconsistency is also present in latest QEMU from git. Why?

IDC scripts: comments containing disassembled Thumb code were updated (DIGIC 6/7/8 and 1300D).